# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Me"
        ]
      }
    },
    "/auth/login": {
      "get": {
        "operationId": "SsoCallbackController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Starts a sign-in: mints a single-use state, keeps it (with the page to\nreturn to) in a cookie only the callback receives, and sends the browser\nto Core Hub's web app, which signs the user in or renews their session.\n\nNo callback_url is sent: Core Hub only ever redirects to the one in its\nregistry, and refuses a different one with 400.",
        "tags": [
          "SsoCallback"
        ]
      }
    },
    "/auth/callback": {
      "get": {
        "operationId": "SsoCallbackController_callback",
        "parameters": [
          {
            "name": "access_token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Bearer",
                "bearer"
              ]
            }
          },
          {
            "name": "expires_in",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 32,
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Where Core Hub sends the browser with a token. Three outcomes (SSO spec D3):\n\n - no `state`: Core Hub started this sign-in itself, e.g. a sidebar click.\n   The token is dropped unused and the browser restarts at /auth/login,\n   which comes back here with a state this subsystem can check. A link\n   carrying an attacker's token therefore signs the victim in as the\n   victim, never as the attacker.\n - a `state` that does not match the cookie: 401, and no redirect - if\n   the browser refuses cookies, redirecting again would loop forever.\n - a matching state and a valid token: the token becomes the session\n   cookie and the browser goes to the page it started from.\n\nThe URL carries the access token, so it is never logged - only the path.",
        "tags": [
          "SsoCallback"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "SsoCallbackController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Signs out everywhere (SSO spec D8). Dropping only this subsystem's cookie\nwould be undone by the next click, which Core Hub - still signed in -\nwould answer with a fresh token. So the browser goes on to Core Hub's\n/logout, where the user confirms ending the Core Hub session.\n\nPublic on purpose: it must work with an expired or missing session, and\nall it does is clear this subsystem's own cookies.",
        "tags": [
          "SsoCallback"
        ]
      }
    },
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/api/v1/profiles/me": {
      "get": {
        "operationId": "ProfilesController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "patch": {
        "operationId": "ProfilesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/v1/tags": {
      "get": {
        "operationId": "TagsController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 30,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      }
    },
    "/api/v1/tag-suggestions": {
      "post": {
        "operationId": "TagsController_suggest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagSuggestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Computes a suggestion; nothing is stored, so it answers 200 rather than 201.",
        "tags": [
          "Tags"
        ]
      }
    },
    "/api/v1/questions": {
      "get": {
        "operationId": "QuestionsController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search over title, body and tag names.",
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 30,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "WAITING",
                "RESOLVED"
              ]
            }
          },
          {
            "name": "mine",
            "required": false,
            "in": "query",
            "description": "Only the caller's own questions.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "unanswered",
            "required": false,
            "in": "query",
            "description": "Waiting questions nobody has commented on yet.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Questions"
        ]
      },
      "post": {
        "operationId": "QuestionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      }
    },
    "/api/v1/questions/{id}": {
      "get": {
        "operationId": "QuestionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      },
      "patch": {
        "operationId": "QuestionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      },
      "delete": {
        "operationId": "QuestionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Questions"
        ]
      }
    },
    "/api/v1/questions/{id}/votes": {
      "post": {
        "operationId": "QuestionsController_vote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      },
      "delete": {
        "operationId": "QuestionsController_unvote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      }
    },
    "/api/v1/questions/{questionId}/comments": {
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/api/v1/comments/{id}": {
      "patch": {
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/api/v1/comments/{id}/votes": {
      "post": {
        "operationId": "CommentsController_vote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_unvote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/api/v1/comments/{id}/verification": {
      "post": {
        "operationId": "CommentsController_verify",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_unverify",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/api/v1/sample-data": {
      "post": {
        "operationId": "SampleDataController_load",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "SampleData"
        ]
      }
    }
  },
  "info": {
    "title": "CSMJU Helpdesk API",
    "description": "กระดานถาม-ตอบ สาขาวิทยาการคอมพิวเตอร์ — ระบบย่อยของ CSMJU2030",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csmju_helpdesk_access_token"
      }
    },
    "schemas": {
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "displayName"
        ]
      },
      "CreateTagSuggestionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 150
          },
          "body": {
            "type": "string",
            "maxLength": 20000
          }
        }
      },
      "CreateQuestionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "body": {
            "type": "string",
            "maxLength": 20000
          },
          "tags": {
            "description": "Omitted or empty = the server suggests tags from the title and body.",
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 60,
              "minLength": 1
            }
          }
        },
        "required": [
          "title",
          "body"
        ]
      },
      "UpdateQuestionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "body": {
            "type": "string",
            "maxLength": 20000
          },
          "tags": {
            "description": "Replaces the whole tag list; an empty list removes every tag.",
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 60,
              "minLength": 1
            }
          }
        }
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 10000
          },
          "parentId": {
            "type": "string",
            "description": "Set = a reply to that answer. Only the question's author may reply.",
            "format": "uuid"
          }
        },
        "required": [
          "body"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 10000
          }
        },
        "required": [
          "body"
        ]
      }
    }
  }
}